Implemented image upload/delete to cloudinary

+++ server/db/index.js
@@ -8,33 +8,49 @@ const options = {
 const pgp = require('pg-promise')(options);
 const  jwt  =  require('jsonwebtoken');
 const bcrypt = require('bcrypt');
+const Datauri = require('datauri')
+const cloudinary = require('cloudinary').v2;
 
 const SECRET_KEY = 'secretkey050440';
 const saltRounds = 10;
 
-console.log('db connection: ', connection.cn)
+// const CLOUDINARY_URL='cloudinary://922781415854575:zcBkJhV1udfKRtYX1dDkZmhw6Yk@dipnvheby'
+cloudinary.config({ 
+  cloud_name: 'dipnvheby', 
+  api_key: '922781415854575', 
+  api_secret: 'zcBkJhV1udfKRtYX1dDkZmhw6Yk' 
+});
+
 const db = pgp(connection.cn); // database instance;
 
-const resizeImage = (path, file) =>  {
-  // Resize to height 432 for card display
-  sharp(path)
-      .resize(648, 432)
-      .toFile('server/api/uploads/images/height_432/' + file, (error, info) => {
-        if (error) {
-          console.error(error)
-          return new Error('Error resizing image (sharp)')
-        }
-      })
-  // Resize to tumbnails 100X100
-  sharp(path)
-      .resize(100, 100)
-      .toFile('server/api/uploads/images/thumbnails/' + file, (error, info) => {
-        if (error) {
-          console.error(error)
-          return new Error('Error resizing image (sharp)')
-        }
-      })
-}
+// const uploadImage = (image) => {
+//   let dUri = new Datauri()
+//   dUri.format(path.extname(image.originalname).toString(),image.buffer);
+//   cloudinary.uploader.upload(dUri.content, (error, result) => {
+//     console.log('cloudinary result: ',result)
+//     console.log('cludinary error: ', error)
+//   });
+// }
+// const resizeImage = (path, file) =>  {
+//   // Resize to height 432 for card display
+//   sharp(path)
+//       .resize(648, 432)
+//       .toFile('server/api/uploads/images/height_432/' + file, (error, info) => {
+//         if (error) {
+//           console.error(error)
+//           return new Error('Error resizing image (sharp)')
+//         }
+//       })
+//   // Resize to tumbnails 100X100
+//   sharp(path)
+//       .resize(100, 100)
+//       .toFile('server/api/uploads/images/thumbnails/' + file, (error, info) => {
+//         if (error) {
+//           console.error(error)
+//           return new Error('Error resizing image (sharp)')
+//         }
+//       })
+// }
 // Users
 const getUsers = (req, res, next) => {
   db.any('SELECT * FROM users', [true])
@@ -178,7 +194,7 @@ const updateUser = (req, res, next) => {
 
 // Posts
 const getPosts = (req, res, next) => {
-  db.any('SELECT posts._id, posts.title, posts.description, posts.content, posts.img_name, posts.img_original_name, posts.imgalt, posts.draft, posts.published, posts.userid, posts.created, posts.updated, users.firstname, users.surname FROM posts, users WHERE posts.userid = users._id', [true])
+  db.any('SELECT posts._id, posts.title, posts.description, posts.content, posts.img_url, posts.img_name, posts.img_original_name, posts.imgalt, posts.draft, posts.published, posts.userid, posts.created, posts.updated, users.firstname, users.surname FROM posts, users WHERE posts.userid = users._id', [true])
     .then(data => {
       return (
         res.status(200).json({
@@ -194,7 +210,7 @@ const getPosts = (req, res, next) => {
 }
 
 const getPost = (req, res, next) => {
-  db.one('SELECT posts._id, posts.title, posts.description, posts.content, posts.img_name, posts.img_original_name, posts.imgalt, posts.draft, posts.published, posts.userid, posts.created, posts.updated, users.firstname, users.surname FROM posts, users WHERE posts._id=${id} AND posts.userid = users._id', req.params)
+  db.one('SELECT posts._id, posts.title, posts.description, posts.content, posts.img_url, posts.img_name, posts.img_original_name, posts.imgalt, posts.draft, posts.published, posts.userid, posts.created, posts.updated, users.firstname, users.surname FROM posts, users WHERE posts._id=${id} AND posts.userid = users._id', req.params)
     .then(data => {
       res.status(200).json({
         status: 'success',
@@ -210,42 +226,98 @@ const getPost = (req, res, next) => {
 
 const addPost = (req, res, next) => {
   if (req.file) {
-    resizeImage(req.file.path, req.file.filename)
-  } else {
-    const file = {filename: '', originalname: ''}
-    req.file = file
-  }
-  db.one('INSERT INTO posts(title, description, content, img_name, img_original_name, imgalt, draft, published, userid) VALUES(${body.title}, ${body.description}, ${body.content}, ${file.filename}, ${file.originalname}, ${body.imgalt}, ${body.draft}, ${body.published}, ${body.userid}) RETURNING _id, created', req)
-    .then((data) => {
-      return (
-        res.status(200).json({
-          status: 'success',
-          post: {id: data._id, created: data.created},
-          file: {name: req.file.filename, originalName: req.file.originalname},
-          message: 'Post added'
-        })
-      )
+    // resizeImage(req.file.path, req.file.filename)
+    let dUri = new Datauri()
+    dUri.format(req.file.originalname, req.file.buffer)
+    cloudinary.uploader.upload(dUri.content, function(error, result) {
+      console.log('cloudinary result', result)
+      if (!error) {
+        req.file.filename = result.public_id
+        req.file.img_url = result.url
+        db.one('INSERT INTO posts(title, description, content, img_url, img_name, img_original_name, imgalt, draft, published, userid) VALUES(${body.title}, ${body.description}, ${body.content}, ${file.img_url}, ${file.filename}, ${file.originalname}, ${body.imgalt}, ${body.draft}, ${body.published}, ${body.userid}) RETURNING _id, created', req)
+          .then((data) => {
+            return (
+              res.status(200).json({
+                status: 'success',
+                post: {id: data._id, created: data.created},
+                file: {url: req.file.img_url, name: req.file.filename, originalName: req.file.originalname},
+                message: 'Post added'
+              })
+            )
+          })
+          .catch(err => {
+            console.error(err)
+            return next(err)
+          });
+      } else {
+        console.log('cludinary error: ', error)
+      }
     })
-    .catch(err => {
-      console.error(err)
-      return next(err)
-    });
+  } else {
+    req.body.img_url = ''
+    req.body.img_name = ''
+    req.body.img_original_name = ''
+    req.body.imgalt = ''
+    db.one('INSERT INTO posts(title, description, content, img_url, img_name, img_original_name, imgalt, draft, published, userid) VALUES(${body.title}, ${body.description}, ${body.content}, ${body.img_url}, ${body.img_name}, ${body.img_original_name}, ${body.imgalt}, ${body.draft}, ${body.published}, ${body.userid}) RETURNING _id, created', req)
+      .then((data) => {
+        return (
+          res.status(200).json({
+            status: 'success',
+            post: {id: data._id, created: data.created},
+            file: { url: req.body.img_url, name: req.body.filename, originalName: req.body.originalname},
+            message: 'Post added'
+          })
+        )
+      })
+      .catch(err => {
+        console.error(err)
+        return next(err)
+      });
+  }
 }
 
 const updatePost = (req, res, next) => {
-  if (req.file) {
-    resizeImage(req.file.path, req.file.filename)
-  } else {
-    const file = {filename: req.body.img_name, originalname: req.body.img_original_name}
-    req.file = file
-  }
   req.body.updated = new Date()
-  db.none('UPDATE posts SET title=${body.title}, description=${body.description}, content=${body.content}, img_name=${file.filename}, img_original_name=${file.originalname}, imgalt=${body.imgalt}, draft=${body.draft}, published=${body.published}, updated=${body.updated} WHERE _id=${params.id}', req)
+  if (req.body.img_name != '' & req.file != null) {
+    // Delete old file on cloudinary
+    cloudinary.uploader.destroy(req.body.img_name, function(result) { 
+      console.log(result)
+    })
+  }         
+  if (req.file) {
+    // Upload new file to cloudinary
+    let dUri = new Datauri()
+    dUri.format(req.file.originalname, req.file.buffer)
+    cloudinary.uploader.upload(dUri.content, function(error, result) {
+      if (!error) {
+        req.file.img_url = result.url
+        req.file.filename = result.public_id
+        db.none('UPDATE posts SET title=${body.title}, description=${body.description}, content=${body.content}, img_url=${file.img_url}, img_name=${file.filename}, img_original_name=${file.originalname}, imgalt=${body.imgalt}, draft=${body.draft}, published=${body.published}, updated=${body.updated} WHERE _id=${params.id}', req)
+          .then(() => {
+            res.status(202).json({
+              status: 'success',
+              post: {updated: req.body.updated},
+              file: {url: req.file.img_url, name: req.file.filename, originalName: req.file.originalname},
+              message: 'Post updated'
+            })
+          })
+          .catch(err => {
+            console.error(err)
+            return next(err)
+          })
+      } else {
+        console.log(error)
+      }
+    })    
+    // resizeImage(req.file.path, req.file.filename)
+  } else {
+    console.log(req.body)
+    db.none('UPDATE posts SET title=${body.title}, description=${body.description}, content=${body.content}, imgalt=${body.imgalt}, draft=${body.draft}, published=${body.published}, updated=${body.updated} WHERE _id=${params.id}', req)
     .then(() => {
       res.status(202).json({
         status: 'success',
         post: {updated: req.body.updated},
-        file: {name: req.file.filename, originalName: req.file.originalname},
+        file: {url: req.body.img_url, name: req.body.img_name, originalName: req.body.img_original_name},
         message: 'Post updated'
       })
     })
@@ -253,20 +325,64 @@ const updatePost = (req, res, next) => {
       console.error(err)
       return next(err)
     })
+  }
 }
 
 const deletePost = (req, res, next) => {
-  db.none('DELETE FROM posts WHERE _id=${id}', req.params)
-    .then(() => {
-      res.status(200).json({
-        status: 'succesful',
-        message: 'Post deleted'
+  console.log(req.query.image)
+  if (req.query.image != '') {
+    // Delete image on cloudinary
+    cloudinary.uploader.destroy(req.query.image, function(result) {
+    db.none('DELETE FROM posts WHERE _id=${id}', req.params)
+      .then(() => {
+        res.status(200).json({
+          status: 'succesful',
+          message: 'Post deleted'
+        })
       })
-    })
-    .catch(err => {
-      console.error(err)
-      return next(err)
-    })
+      .catch(err => {
+        console.error(err)
+        return next(err)
+      })
+  })
+  } else {
+    db.none('DELETE FROM posts WHERE _id=${id}', req.params)
+      .then(() => {
+        res.status(200).json({
+          status: 'succesful',
+          message: 'Post deleted'
+        })
+      })
+      .catch(err => {
+        console.error(err)
+        return next(err)
+      })
+  }
+}
+
+const removeImage = (req, res, next) => {
+  console.log(req.body)
+  // Delete image on cloudinary
+  cloudinary.uploader.destroy(req.body.image, function(result) {
+    const image = {
+      img_url: '',
+      img_name: '',
+      img_original_name: '',
+      imgalt: ''
+    }
+    req.image = image
+    db.none('UPDATE posts SET img_url=${image.img_url}, img_name=${image.img_name}, img_original_name=${image.img_original_name}, imgalt=${image.imgalt} WHERE _id=${params.id}', req)
+      .then(() => {
+        res.status(200).json({
+          status: 'succesful',
+          message: 'Image deleted'
+        })
+      })
+      .catch(err => {
+        console.error(err)
+        return next(err)
+      })
+  })
 }
 
 module.exports = {
@@ -280,4 +396,5 @@ module.exports = {
   addPost,
   updatePost,
   deletePost,
+  removeImage
 }
